change saving answer on testset

+++ commons/model.py
@@ -43,7 +43,7 @@ def _get_pretrained_model(
         
         elif distribution_type == "ddp":
             quantization_config=None
-            loading_dtype = torch.float16
+            loading_dtype = torch.float32
         
         else:
             raise NotImplementedError
